+++ ethereum/core/src/main/java/org/hyperledger/besu/ethereum/transaction/BlockSimulator.java
@@ -120,8 +120,8 @@ public class BlockSimulator {
       final long rpcGasCap) {
     this.worldStateArchive = worldStateArchive;
     this.protocolSchedule = protocolSchedule;
-    this.miningConfiguration = miningConfiguration;
     this.transactionSimulator = transactionSimulator;
+    this.miningConfiguration = miningConfiguration;
     this.blockchain = blockchain;
     this.rpcGasCap = rpcGasCap;
   }
@@ -207,6 +207,7 @@ public class BlockSimulator {
               resolveValidationParams(simulationParameter),
               simulationParameter.isTraceTransfers(),
               simulationParameter.isReturnTrieLog(),
+              simulationParameter.isEnforceConsensusGasLimit(),
               simulationParameter::getFakeSignature,
               blockHashCache,
               simulationCumulativeGasUsed,
@@ -225,7 +226,7 @@ public class BlockSimulator {
     if (!simulationParameter.isValidation()) {
       return NON_STRICT_PARAMS;
     }
-    return simulationParameter.isEnforceConsensusGasLimitCaps()
+    return simulationParameter.isEnforceConsensusGasLimit()
         ? CONSENSUS_STRICT_VALIDATION_PARAMS
         : STRICT_VALIDATION_PARAMS;
   }
@@ -246,6 +247,7 @@ public class BlockSimulator {
       final TransactionValidationParams validationParams,
       final boolean isTraceTransfers,
       final boolean returnTrieLog,
+      final boolean enforceConsensusGasLimit,
       final Supplier<SECPSignature> signatureSupplier,
       final Map<Long, Hash> blockHashCache,
       final long simulationCumulativeGasUsed,
@@ -267,7 +269,12 @@ public class BlockSimulator {
     ProtocolSpec protocolSpec = protocolSchedule.getByBlockHeader(syntheticNextBlockHeader);
 
     BlockHeader overridenBaseBlockHeader =
-        overrideBlockHeader(baseBlockHeader, protocolSpec, blockOverrides, shouldValidate);
+        overrideBlockHeader(
+            baseBlockHeader,
+            protocolSpec,
+            blockOverrides,
+            shouldValidate,
+            enforceConsensusGasLimit);
 
     blockStateCall
         .getStateOverrideMap()
@@ -659,7 +666,8 @@ public class BlockSimulator {
       final BlockHeader header,
       final ProtocolSpec newProtocolSpec,
       final BlockOverrides blockOverrides,
-      final boolean shouldValidate) {
+      final boolean shouldValidate,
+      final boolean enforceConsensusGasLimit) {
     long timestamp = blockOverrides.getTimestamp().orElseThrow();
     long blockNumber = blockOverrides.getBlockNumber().orElseThrow();
 
@@ -678,7 +686,11 @@ public class BlockSimulator {
             .gasLimit(
                 blockOverrides
                     .getGasLimit()
-                    .orElseGet(() -> getNextGasLimit(newProtocolSpec, header, blockNumber)))
+                    .orElseGet(
+                        () ->
+                            enforceConsensusGasLimit
+                                ? getNextGasLimit(newProtocolSpec, header, blockNumber)
+                                : header.getGasLimit()))
             .extraData(blockOverrides.getExtraData().orElse(Bytes.EMPTY))
             .prevRandao(blockOverrides.getMixHashOrPrevRandao().orElse(Bytes32.ZERO));